moved tolua from util to core because of circular dependencies reverted some bugfixes in ParticleInterface

+++ src/orxonox/objects/SpaceShip.h
@@ -81,7 +81,7 @@ namespace orxonox
             float mouseX_;
             float mouseY_;
 
-            int emitterRate_;
+            float emitterRate_;
     };
 }
